CLAB-44 refact: validateRefreshToken 메서드화
