--- src/main/java/com/cpplab/domain/auth/controller/ReissueController.java
@@ -1,40 +1,0 @@
-package com.cpplab.domain.auth.controller;
-
-
-import com.cpplab.domain.auth.service.ReissueService;
-import jakarta.servlet.http.Cookie;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class ReissueController {
-
-    private final ReissueService reissueService;
-
-    public ReissueController(ReissueService reissueService) {
-        this.reissueService = reissueService;
-    }
-
-    @PostMapping("/reissue")
-    public ResponseEntity<?> reissue(HttpServletRequest request, HttpServletResponse response) {
-
-        //get refresh token
-        String refresh = null;
-        Cookie[] cookies = request.getCookies();
-        for (Cookie cookie : cookies) {
-            if (cookie.getName().equals("refresh")) {
-                refresh = cookie.getValue();
-            }
-        }
-
-        if (refresh == null) {
-            //response status code
-            return new ResponseEntity<>("refresh token null", HttpStatus.BAD_REQUEST);
-        }
-        return reissueService.reissueTokens(refresh, response);
-    }
-}
